+++ experimental/builder/include/ck_tile/builder/testing/conv/fwd_ck.hpp
@@ -3,9 +3,9 @@
 
 #pragma once
 
-#include "ck_tile/builder/testing/conv_fwd.hpp"
-#include "ck_tile/host/kernel_launch.hpp"
+#include "ck_tile/builder/testing/conv/fwd.hpp"
 #include "ck_tile/builder/factory/helpers/ck/conv_elementwise_op.hpp"
+#include "ck_tile/host/kernel_launch.hpp"
 #include <type_traits>
 #include <array>
