+++ profiler/include/profiler/grouped_convolution_forward_tile_algs.hpp
@@ -9,8 +9,8 @@
 #include "grouped_convolution_signatures.hpp"
 
 #include "ck_tile/builder/testing/filter_extent.hpp"
-#include "ck_tile/builder/testing/conv_fwd_ck_tile.hpp"
-#include "ck_tile/builder/testing/conv_fwd_reference.hpp"
+#include "ck_tile/builder/testing/conv/fwd_ck_tile.hpp"
+#include "ck_tile/builder/testing/conv/reference.hpp"
 #include "ck_tile/builder/conv_builder.hpp"
 
 namespace ck_tile::builder::profiling {
