+++ profiler/src/profile_grouped_conv_fwd_tile.cpp
@@ -6,7 +6,7 @@
 #include <initializer_list>
 #include <cstdlib>
 
-#include "ck_tile/builder/testing/conv_fwd_ck_tile.hpp"
+#include "ck_tile/builder/testing/conv/fwd_ck_tile.hpp"
 #include "ck_tile/host/device_prop.hpp"
 #include "profiler/grouped_convolution_forward_tile_algs.hpp"
